+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 development:
   * add SQL logger to development environment
+  * generate public dir when running generator
 
 0.3.0: